+++ lizard_map/static/lizard_map/jquery.workspace.js
@@ -93,7 +93,7 @@ function show_popup(data) {
                 $("#movable-dialog-content").html(data.html[0]);
             } else {
                 // Build up html with tabs.
-                html = '<ul class="nav nav-tabs css-tabs">';
+                html = '<ul id="popup-tabs" class="nav nav-tabs css-tabs">';
                 for (i = 0; i < data.html.length; i += 1) {
                     html += '<li><a href="#">Resultaat ';
                     html += (i + 1) + '</a></li>';
@@ -107,9 +107,11 @@ function show_popup(data) {
                 html += '</div>';
 
                 $("#movable-dialog-content").html(html);
-                $(".nav-tabs").tabs("div.popup-panes > div.pane",
+                $("#popup-tabs").tabs("div.popup-panes > div.pane",
                                 {'effect': 'map_popup'});
             }
+            $("#popup-subtabs").tabs("div.popup-subpanes > div.subpane",
+                            {'effect': 'map_popup'});
             //dialogOverlay();
             $("#movable-dialog").dialog("open");
             //if (data.html.length === 1) {