+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog of lizard-map
 4.6 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Added popup with subtabs.
 
 
 4.5 (2012-08-14)